+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 19. Moved `GrailLogo` and `backgroundImage` (used for `SignInPage`) to new package `@grail/common-private`.
     When upgrading, you will need to specify `logo` and `backgroundImage` to `NavBar` and `SignInPage` as appropriate.
 20. Send `{ isUserSearchAction: true }` when a user changes a value in `OmniSearchBar`.
+21. Add static footer to `PagedTable` for non-paginated tables, showing number of records.
 
 ## 1.0.0 - 2019-06-17